Modified project structure
