--- src/main/java/com/codingshadows/auth/AuthApplication.java
@@ -1,13 +1,0 @@
-package com.codingshadows.auth;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AuthApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AuthApplication.class, args);
-	}
-
-}
