--- src/main/java/com/codingshadows/auth/controller/HelloResource.java
@@ -1,14 +1,0 @@
-package com.codingshadows.auth.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class HelloResource {
-    @RequestMapping({"hello"})
-    public String firstPage() {
-        return "Hello World";
-    }
-}
