--- src/main/java/com/codingshadows/auth/exception/TokenRefreshException.java
@@ -1,15 +1,0 @@
-package com.codingshadows.auth.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-import java.io.Serial;
-
-@ResponseStatus(HttpStatus.FORBIDDEN)
-public class TokenRefreshException extends RuntimeException {
-    @Serial
-    private static final long serialVersionUID = 1L;
-    public TokenRefreshException(String token, String message) {
-        super(String.format("Failed for [%s]: %s", token, message));
-    }
-}
