--- src/main/java/com/codingshadows/auth/filters/JwtRequestFilter.java
@@ -1,59 +1,0 @@
-package com.codingshadows.auth.filters;
-
-import com.codingshadows.auth.advice.UserControllerAdvice;
-import com.codingshadows.auth.exception.JwtException;
-import com.codingshadows.auth.model.ErrorMessage;
-import com.codingshadows.auth.service.UserDetailsServiceImpl;
-import com.codingshadows.auth.util.JwtUtil;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@Component
-@RequiredArgsConstructor
-public class JwtRequestFilter extends OncePerRequestFilter {
-    private final UserDetailsServiceImpl userDetailsService;
-    private final JwtUtil jwtUtil;
-
-    private final UserControllerAdvice userControllerAdvice;
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain chain) throws ServletException, IOException {
-        final String authorizationHeader = request.getHeader("Authorization");
-        String username = null;
-        String jwt = null;
-
-        if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
-            jwt = authorizationHeader.substring(7);
-            username = jwtUtil.extractEmail(jwt);
-        }
-
-        if (username != null && SecurityContextHolder.getContext().getAuthentication() == null) {
-            UserDetails userDetails = this.userDetailsService.loadUserByUsername(username);
-            try {
-                if (jwtUtil.validateToken(jwt, userDetails)) {
-                    UsernamePasswordAuthenticationToken usernamePasswordAuthenticationToken = new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
-                    usernamePasswordAuthenticationToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
-                    SecurityContextHolder.getContext().setAuthentication(usernamePasswordAuthenticationToken);
-                }
-            } catch (JwtException e) {
-                ErrorMessage errorMessage = userControllerAdvice.handleJWTExceptionFromFilter(e);
-                response.setStatus(errorMessage.getStatusCode());
-                response.getWriter().write(errorMessage.getMessage());
-                return;
-            }
-        }
-        chain.doFilter(request, response);
-    }
-
-}
