--- src/main/java/com/codingshadows/auth/model/ErrorMessage.java
@@ -1,15 +1,0 @@
-package com.codingshadows.auth.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-import java.util.Date;
-
-@Data
-@AllArgsConstructor
-public class ErrorMessage {
-    private int statusCode;
-    private Date timestamp;
-    private String message;
-    private String description;
-}
