--- src/main/java/com/codingshadows/auth/model/RefreshToken.java
@@ -1,16 +1,0 @@
-package com.codingshadows.auth.model;
-
-import lombok.Data;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-import java.time.Instant;
-
-@Data
-@Document
-public class RefreshToken {
-    @Id
-    private String token;
-    private String userId;
-    private Instant expiryDate;
-}
