--- src/main/java/com/codingshadows/auth/model/User.java
@@ -1,56 +1,0 @@
-package com.codingshadows.auth.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.ArrayList;
-import java.util.Collection;
-
-@Document(collection = "users")
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class User implements UserDetails {
-    @Id
-    private String userID;
-    private String username;
-    private String password;
-    private String email;
-    private String firstName;
-    private String lastName;
-    private Boolean enabled;
-    private Boolean accountNonExpired;
-    private Boolean accountNonLocked;
-    private boolean credentialsNonExpired;
-    private Collection<? extends GrantedAuthority> authorities = new ArrayList<>();
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return authorities;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return accountNonExpired;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return accountNonLocked;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return credentialsNonExpired;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return enabled;
-    }
-}
