--- src/main/java/com/codingshadows/auth/payload/request/AuthenticationRequest.java
@@ -1,15 +1,0 @@
-package com.codingshadows.auth.payload.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthenticationRequest implements Serializable {
-    private String email;
-    private String password;
-}
