--- src/main/java/com/codingshadows/auth/payload/request/LogOutRequest.java
@@ -1,12 +1,0 @@
-package com.codingshadows.auth.payload.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class LogOutRequest {
-    private String userId;
-}
