--- src/main/java/com/codingshadows/auth/payload/request/TokenRefreshRequest.java
@@ -1,15 +1,0 @@
-package com.codingshadows.auth.payload.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotBlank;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class TokenRefreshRequest {
-    @NotBlank
-    private String refreshToken;
-}
