--- src/main/java/com/codingshadows/auth/payload/response/AuthenticationResponse.java
@@ -1,6 +1,0 @@
-package com.codingshadows.auth.payload.response;
-
-import com.codingshadows.auth.model.RefreshToken;
-
-public record AuthenticationResponse(String jwt, RefreshToken refreshToken) {
-}
