--- src/main/java/com/codingshadows/auth/payload/response/MessageResponse.java
@@ -1,12 +1,0 @@
-package com.codingshadows.auth.payload.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class MessageResponse {
-    private String message;
-}
