--- src/main/java/com/codingshadows/auth/payload/response/TokenRefreshResponse.java
@@ -1,14 +1,0 @@
-package com.codingshadows.auth.payload.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class TokenRefreshResponse {
-    private String accessToken;
-    private String refreshToken;
-    private final String tokenType = "Bearer";
-}
