--- src/main/java/com/codingshadows/auth/repository/RefreshTokenRepository.java
@@ -1,8 +1,0 @@
-package com.codingshadows.auth.repository;
-
-import com.codingshadows.auth.model.RefreshToken;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-public interface RefreshTokenRepository extends MongoRepository<RefreshToken, String> {
-    void deleteByUserId(String userId);
-}
