--- src/main/java/com/codingshadows/auth/service/RefreshTokenService.java
@@ -1,45 +1,0 @@
-package com.codingshadows.auth.service;
-
-import com.codingshadows.auth.exception.TokenRefreshException;
-import com.codingshadows.auth.model.RefreshToken;
-import com.codingshadows.auth.repository.RefreshTokenRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-
-import java.time.Instant;
-import java.util.Optional;
-import java.util.UUID;
-
-@Service
-@RequiredArgsConstructor
-public class RefreshTokenService {
-    @Value("${codingshadows.app.jwtRefreshTokenExpirationMs}")
-    private Long jwtRefreshTokenExpirationMs;
-    private final RefreshTokenRepository refreshTokenRepository;
-
-    public Optional<RefreshToken> findByToken(String token) {
-        return refreshTokenRepository.findById(token);
-    }
-
-    public RefreshToken createRefreshToken(String userId) {
-        RefreshToken refreshToken = new RefreshToken();
-        refreshToken.setUserId(userId);
-        refreshToken.setExpiryDate(Instant.now().plusMillis(jwtRefreshTokenExpirationMs));
-        refreshToken.setToken(UUID.randomUUID().toString());
-        refreshToken = refreshTokenRepository.save(refreshToken);
-        return refreshToken;
-    }
-
-    public RefreshToken verifyExpiration(RefreshToken token) {
-        if (token.getExpiryDate().compareTo(Instant.now()) < 0) {
-            refreshTokenRepository.delete(token);
-            throw new TokenRefreshException(token.getToken(), "Refresh token was expired. Please make a new signin request");
-        }
-        return token;
-    }
-
-    public void deleteByUserId(String userId) {
-        refreshTokenRepository.deleteByUserId(userId);
-    }
-}
