--- src/main/java/com/codingshadows/auth/service/UserDetailsServiceImpl.java
@@ -1,26 +1,0 @@
-package com.codingshadows.auth.service;
-
-import com.codingshadows.auth.repository.UserRepository;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserDetailsServiceImpl implements UserDetailsService {
-    final
-    UserRepository userRepository;
-
-    public UserDetailsServiceImpl(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return userRepository.findUserByUsername(username);
-    }
-
-    public UserDetails loadUserByEmail(String email) throws UsernameNotFoundException {
-        return userRepository.findUserByEmail(email);
-    }
-}
