+++ src/main/java/com/dbproject/cvapp/advice/UserControllerAdvice.java
@@ -1,0 +1,33 @@
+package com.dbproject.cvapp.advice;
+
+import com.dbproject.cvapp.exception.JwtException;
+import com.dbproject.cvapp.model.ErrorMessage;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.context.request.WebRequest;
+
+import java.util.Date;
+
+@RestControllerAdvice
+public class UserControllerAdvice {
+    @ExceptionHandler(value = JwtException.class)
+    @ResponseStatus(HttpStatus.FORBIDDEN)
+    public ErrorMessage handleJWTException(JwtException ex, WebRequest request) {
+        return new ErrorMessage(
+                HttpStatus.FORBIDDEN.value(),
+                new Date(),
+                ex.getMessage(),
+                request.getDescription(false));
+    }
+
+    @ResponseStatus(HttpStatus.FORBIDDEN)
+    public ErrorMessage handleJWTExceptionFromFilter(JwtException ex) {
+        return new ErrorMessage(
+                HttpStatus.FORBIDDEN.value(),
+                new Date(),
+                ex.getMessage(),
+                ex.getMessage());
+    }
+}
