+++ src/main/java/com/dbproject/cvapp/controller/HelloResource.java
@@ -1,0 +1,14 @@
+package com.dbproject.cvapp.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class HelloResource {
+    @RequestMapping({"hello"})
+    public String firstPage() {
+        return "Hello World";
+    }
+}
