+++ src/main/java/com/dbproject/cvapp/model/ErrorMessage.java
@@ -1,0 +1,15 @@
+package com.dbproject.cvapp.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+import java.util.Date;
+
+@Data
+@AllArgsConstructor
+public class ErrorMessage {
+    private int statusCode;
+    private Date timestamp;
+    private String message;
+    private String description;
+}
