+++ src/main/java/com/dbproject/cvapp/payload/request/AuthenticationRequest.java
@@ -1,0 +1,15 @@
+package com.dbproject.cvapp.payload.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.io.Serializable;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class AuthenticationRequest implements Serializable {
+    private String email;
+    private String password;
+}
