+++ src/main/java/com/dbproject/cvapp/payload/request/LogOutRequest.java
@@ -1,0 +1,12 @@
+package com.dbproject.cvapp.payload.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class LogOutRequest {
+    private String userId;
+}
