+++ src/main/java/com/dbproject/cvapp/payload/request/TokenRefreshRequest.java
@@ -1,0 +1,15 @@
+package com.dbproject.cvapp.payload.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.validation.constraints.NotBlank;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class TokenRefreshRequest {
+    @NotBlank
+    private String refreshToken;
+}
