+++ src/main/java/com/dbproject/cvapp/payload/response/AuthenticationResponse.java
@@ -1,0 +1,6 @@
+package com.dbproject.cvapp.payload.response;
+
+import com.dbproject.cvapp.model.RefreshToken;
+
+public record AuthenticationResponse(String jwt, RefreshToken refreshToken) {
+}
