+++ src/main/java/com/dbproject/cvapp/payload/response/MessageResponse.java
@@ -1,0 +1,12 @@
+package com.dbproject.cvapp.payload.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class MessageResponse {
+    private String message;
+}
