+++ src/main/java/com/dbproject/cvapp/payload/response/TokenRefreshResponse.java
@@ -1,0 +1,14 @@
+package com.dbproject.cvapp.payload.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class TokenRefreshResponse {
+    private String accessToken;
+    private String refreshToken;
+    private final String tokenType = "Bearer";
+}
