+++ src/main/java/com/dbproject/cvapp/repository/RefreshTokenRepository.java
@@ -1,0 +1,8 @@
+package com.dbproject.cvapp.repository;
+
+import com.dbproject.cvapp.model.RefreshToken;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+public interface RefreshTokenRepository extends MongoRepository<RefreshToken, String> {
+    void deleteByUserId(String userId);
+}
