+++ src/main/java/com/dbproject/cvapp/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.dbproject.cvapp.repository;
+
+import com.dbproject.cvapp.model.User;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+public interface UserRepository extends MongoRepository<User, String> {
+    User findUserByUsername(String username);
+
+    User findUserByEmail(String email);
+
+    User findUserByUserID(String userID);
+}
