--- src/test/java/com/codingshadows/auth/AuthApplicationTests.java
@@ -1,13 +1,0 @@
-package com.codingshadows.auth;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AuthApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
