+++ src/test/java/com/dbproject/cvapp/AuthApplicationTests.java
@@ -1,0 +1,13 @@
+package com.dbproject.cvapp;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AuthApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
